add mongodb support to app
