+++ generator/requirements.txt
@@ -1,3 +1,3 @@
 base-62
 Flask
-validators
+pymongo
